analytics/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.38
botocore==1.42.38
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.129.0
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.16.4
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
openai==2.16.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-transformers==1.1.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==2.2.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.35.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
